added conftest.py, moved an fixture to conftest.py
